# metadata: Use strings.EqualFold for ValueFromIncomingContext