--- a/metadata/metadata_test.go
+++ b/metadata/metadata_test.go
@@ -198,6 +198,40 @@ func (s) TestDelete(t *testing.T) {
 	}
 }
 
+func (s) TestFromIncomingContext(t *testing.T) {
+	md := Pairs(
+		"X-My-Header-1", "42",
+	)
+	ctx := NewIncomingContext(context.Background(), md)
+
+	// Check that we lowercase if callers modify md after NewIncomingContext
+	md["X-INCORRECT-UPPERCASE"] = []string{"foo"}
+
+	result, found := FromIncomingContext(ctx)
+	if !found {
+		t.Fatal("FromIncomingContext must return metadata")
+	}
+	expected := MD{
+		"x-my-header-1":         []string{"42"},
+		"x-incorrect-uppercase": []string{"foo"},
+	}
+	if !reflect.DeepEqual(result, expected) {
+		t.Errorf("FromIncomingContext returned %#v, expected %#v", result, expected)
+	}
+
+	// ensure modifying result does not modify the value in the context
+	result["new_key"] = []string{"foo"}
+	result["x-my-header-1"][0] = "mutated"
+
+	result2, found := FromIncomingContext(ctx)
+	if !found {
+		t.Fatal("FromIncomingContext must return metadata")
+	}
+	if !reflect.DeepEqual(result2, expected) {
+		t.Errorf("FromIncomingContext after modifications returned %#v, expected %#v", result2, expected)
+	}
+}
+
 func (s) TestValueFromIncomingContext(t *testing.T) {
 	md := Pairs(
 		"X-My-Header-1", "42",
@@ -206,6 +240,9 @@ func (s) TestValueFromIncomingContext(t *testing.T) {
 		"x-my-header-3", "44",
 	)
 	ctx := NewIncomingContext(context.Background(), md)
+
+	// Check that we lowercase if callers modify md after NewIncomingContext
+	md["X-INCORRECT-UPPERCASE"] = []string{"foo"}
 
 	for _, test := range []struct {
 		key  string
@@ -226,6 +263,10 @@ func (s) TestValueFromIncomingContext(t *testing.T) {
 		{
 			key:  "x-unknown",
 			want: nil,
+		},
+		{
+			key:  "x-incorrect-uppercase",
+			want: []string{"foo"},
 		},
 	} {
 		v := ValueFromIncomingContext(ctx, test.key)
@@ -348,8 +389,22 @@ func BenchmarkFromIncomingContext(b *testing.B) {
 func BenchmarkValueFromIncomingContext(b *testing.B) {
 	md := Pairs("X-My-Header-1", "42")
 	ctx := NewIncomingContext(context.Background(), md)
-	b.ResetTimer()
-	for n := 0; n < b.N; n++ {
-		ValueFromIncomingContext(ctx, "x-my-header-1")
-	}
+
+	b.Run("key-found", func(b *testing.B) {
+		for n := 0; n < b.N; n++ {
+			result := ValueFromIncomingContext(ctx, "x-my-header-1")
+			if len(result) != 1 {
+				b.Fatal("ensures not optimized away")
+			}
+		}
+	})
+
+	b.Run("key-not-found", func(b *testing.B) {
+		for n := 0; n < b.N; n++ {
+			result := ValueFromIncomingContext(ctx, "key-not-found")
+			if len(result) != 0 {
+				b.Fatal("ensures not optimized away")
+			}
+		}
+	})
 }