@evanj 2023-10-19
I may have misunderstood what the comments that I removed mean. I noticed this because I was looking at profiles for a program that does a lot of tiny gRPC requests, and noticed with our various libraries, we call `metadata.FromIncomingContext` 3 times for each request, for a total of about 2.5% of all CPU consumed by this service. It would be nice if this could just return an immutable object instead.

I am very happy to see ValueFromIncomingContext instead! It should be a nice easy improvement across all our services. Although it is a bit unfortunate that the values are in a map that has to be linearly searched on a "not found" request.