@evanj 2023-10-21
Thank you for the detailed discussion! I suspect that the ~3.5% CPU overhead I'm seeing is going to be greatly reduced when I change our internal helpers to use `ValueFromIncomingContext`. Unfortunately, Datadog's Go tracing library needs to iterate over all keys, so it can't use this. I'll try to find an hour next week to take a more careful look at this API. I'll file a separate issue if I think there is a "reasonable" way to evolve it in some way that might be able to avoid cloning the map on each call. Maybe a new call that returns an immutable view wrapper around the MD?

[1] https://github.com/DataDog/dd-trace-go/blob/00055c7bdeed864cd0008ba1216c3b508dce0874/contrib/google.golang.org/grpc/grpc.go#L69