@dfawley 2023-10-24
> Maybe a new call that returns an immutable view wrapper around the MD?

Something like that might be a good idea and not require a total rework of the metadata package (or a v2).  Thanks!